+++ packages/wiz-ui/src/components/base/inputs/date-picker/date-picker.vue
@@ -1,6 +1,7 @@
 <template>
   <WizPopupContainer>
     <button
+      type="button"
       :class="[
         datePickerStyle,
         datePickerVariantStyle[variant],
@@ -26,6 +27,7 @@
           <WizIcon size="xl2" color="gray.500" :icon="WizICalendar" />
         </span>
         <button
+          type="button"
           v-else
           :class="datePickerCancelButtonStyle"
           :aria-label="ARIA_LABELS.DATE_PICKER_CANCEL"
@@ -53,6 +55,7 @@
             </WizText>
             <WizVStack>
               <button
+                type="button"
                 :aria-label="ARIA_LABELS.YEAR_SELECTOR_NEXT"
                 :class="[datePickerYearSelectorItemStyle]"
                 @click="clickToNextYear"
@@ -66,6 +69,7 @@
                 />
               </button>
               <button
+                type="button"
                 :aria-label="ARIA_LABELS.YEAR_SELECTOR_PREV"
                 :class="[datePickerYearSelectorItemStyle]"
                 @click="clickToPreviousYear"
@@ -85,6 +89,7 @@
           </WizText>
           <div :class="datePickerMonthSelectorStyle">
             <button
+              type="button"
               :aria-label="ARIA_LABELS.MONTH_SELECTOR_PREV"
               :class="datePickerMonthSelectorItemStyle"
               @click="clickToPreviousMonth"
@@ -92,6 +97,7 @@
               <WizIcon size="md" color="inherit" :icon="WizIChevronLeft" />
             </button>
             <button
+              type="button"
               :aria-label="ARIA_LABELS.MONTH_SELECTOR_NEXT"
               :class="datePickerMonthSelectorItemStyle"
               @click="clickToNextMonth"
@@ -111,7 +117,7 @@
                 ]
               : []
           "
-          @click="(date) => (calendarValue = date)"
+          @click="handleClickCalendar"
           :currentMonth="currentMonth"
           filledWeeks
           :disabledDate="disabledDate"
@@ -289,4 +295,6 @@ const variant = computed(() => {
   if (calendarValue.value) return "selected";
   return "default";
 });
+
+const handleClickCalendar = (date: Date) => (calendarValue.value = date);
 </script>